Problem with the print:
Located in Typer with Pi. I think because the toNames method gets a ctx
that is local to the Pi you're working on, accordingly, it can re-use
the same name. Unsure if it is even a bug or solvable. Especially since
it mainly uses methods that we haven't made ourselves. To cause more
different names, I do add the "_" name to the ctx, so that certainly
doesn't get reused.